+++ selfdrive/car/card.py
@@ -212,8 +212,6 @@ class Car:
     CS.vCruiseCluster = float(v_cruise_cluster_kph)
     CS.softHoldActive = self.v_cruise_helper._soft_hold_active
     CS.activateCruise = self.v_cruise_helper._activate_cruise
-    CS.latEnabled = self.v_cruise_helper._lat_enabled
-    CS.latOverride = self.v_cruise_helper._lat_override
     CS.useLaneLineSpeed = self.v_cruise_helper.useLaneLineSpeedApply
     CS.carrotCruise = 1 if self.v_cruise_helper.carrot_cruise_active else 0
